GUIsupport: Change type of TableWidget value
The current implementation of the return value of the getValue of the
the TableWidget was a numpy array. But this was not so helpful, since
the numpy array type is not part of the nicos standard namespace.

Converting it to a list type makes it applicable in the standard daemon
namespace.

Change-Id: I1f3e714cf1f1f8948bbbf163d7b1f1d21b685da4
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/37794
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>