MLZ/BioDiff: Create a wavelength changer device
Since the selection of the wavelength at the instrument is done with the
help of 2 devices a monochromator crystal followed by a selector device
to remove the higher order reflections, a more complex device has to be
created.
The selector is mounted a the 2theta arm of the monochromator and it
isn't allowed to move this around the monochromator table until a
certain (or minimum) rotation speed is reached.
After adjusting the monochromator crystal to the desired wavelength the
selector itself has to be adjusted to this wavelength as well.

Change-Id: Ibd951bc93f5f1be129899ac2fd6144c814d1d886
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/37808
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>