nicos_demo: Add ProcessCPULoad device
This is a device inspired from nicos_mlz/refsans/psutil, but rewritten.
It monitors the cpu load of a process with the specified pollinterval.
The process is matched from list of strings and compared with cmdline
strings the process was started with.

Example of how to instantiate in a setup file:
histogram_south_cpu = device('nicos_demo.demo.devices.cpuload.ProcessCPULoad',
    description = 'CPU utilization',
    process_strings = ['python', 'poldi_south.ini'],
    pollinterval = 6.0,
    unit = 'percent',
    fmtstr = '%d',
),

Change-Id: I3c264aeb767a6412475ad54269dac7ba010996c7
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/37817
Reviewed-by: Konstantin Kholostov <k.kholostov@fz-juelich.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Stefan Mathis <stefan.mathis@psi.ch>
Reviewed-by: Alexander Söderqvist <alexander.soederqvist@psi.ch>
Reviewed-by: Edward Wall <edward.wall@psi.ch>