SECoP: rebuild cache after a succesful reconnect
When a reconnect does not succeed within about 10 s, the cache of
all parameters of the SECoP devices are cleared. This is a feature.
However, the cache is not rebuilt when reconnected, and this leads
the GUI device panel to not work properly after a reconnection.

This fix does:
- clear only the cache entries of the SECoP parameters
- rebuild the cache for these after a successful reconnect

Fixes #5051

Change-Id: I3d2f1916828513d63586d1a523c6266cb885faad
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/37856
Reviewed-by: Georg Brandl <g.brandl@fz-juelich.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Markus Zolliker <markus.zolliker@psi.ch>