Stressi: Add special offcentered slit
There is some hardware which is using 4 devices to move the center and
the opening. The center movement is done by 2 motors to translate the
whole slit in horizontal resp. vertical direction. The opening is done
by 2 other motors moving the blades pairwise in horizontal or vertical
direction to achieve the opening distance between the blades.

Change-Id: I449854de12e5c4e41d7300c297a213dbee811386
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/37920
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>