FirePOD: Add special Cmdlets for GUI on mobile device
The instrument scientist(s) and engineer(s) wanted to have a special GUI
for a mobile device (tablet), which should be used inside the
experimental area for alignment tasks.

There is a cleaned (in functionality limited) GUI required. Not all of
the Cmdlets should be visible and some of them should have access only
to a limited (configurable) list of devices.

Change-Id: I13c347f5008f10e56c70ff08035cac5ed4a7ba9b
Reviewed-on: https://forge.frm2.tum.de/review/c/frm2/nicos/nicos/+/38093
Tested-by: Jenkins Automated Tests <pedersen+jenkins@frm2.tum.de>
Reviewed-by: Jens Krueger <jens.krueger@frm2.tum.de>